+++ pop-miners/altchain-pop-miner/src/main/kotlin/org/veriblock/lite/core/FullBlock.kt
@@ -25,7 +25,7 @@ class FullBlock(
     merkleRoot: Sha256Hash,
     timestamp: Int,
     difficulty: Int,
-    nonce: Int,
+    nonce: Long,
     val normalTransactions: List<VeriBlockTransaction>,
     val popTransactions: List<VeriBlockPopTransaction>,
     val metaPackage: BlockMetaPackage
